CRUD services for Item and Store

+++ create_db.sql
@@ -1,7 +1,7 @@
 CREATE SEQUENCE stores_sequence START 1 INCREMENT 1;
 
 CREATE TABLE stores (
-  id integer NOT NULL DEFAULT nextval('stores_sequence'),
+  id bigint NOT NULL DEFAULT nextval('stores_sequence'),
   city varchar(100),
   address text
 );
@@ -10,7 +10,7 @@ CREATE TABLE stores (
 CREATE SEQUENCE items_sequence START 1 INCREMENT 1;
 
 CREATE TABLE items(
-  id integer NOT NULL DEFAULT nextval('items_sequence'),
+  id bigint NOT NULL DEFAULT nextval('items_sequence'),
   serial_number varchar(100),
   name varchar(100),
   description text
@@ -20,7 +20,7 @@ CREATE TABLE items(
 CREATE SEQUENCE items_location_sequence START 1 INCREMENT 1;
 
 CREATE TABLE items_location(
-  id integer NOT NULL DEFAULT nextval('items_location_sequence'),
+  id bigint NOT NULL DEFAULT nextval('items_location_sequence'),
   store_id integer,
   item_id integer
 );
@@ -29,8 +29,8 @@ CREATE TABLE items_location(
 CREATE SEQUENCE reviews_sequence START 1 INCREMENT 1;
 
 CREATE TABLE reviews(
-  id integer NOT NULL DEFAULT nextval('reviews_sequence'),
-  item_id integer,
+  id bigint NOT NULL DEFAULT nextval('reviews_sequence'),
+  item_id bigint,
   nr_stars int,
   comment text
 );
